+++ R/assertions.R
@@ -40,11 +40,11 @@ is_count <- function(x, null_ok = FALSE) {
   is.numeric(x) && length(x) == 1L && is.finite(x) && as.integer(x) == x && x > 0L
 }
 
-is_number <- function(x, null_ok = FALSE) {
+is_number <- function(x, lower = -Inf, upper = Inf, null_ok = FALSE) {
   if (null_ok && is.null(x)) {
     return(TRUE)
   }
-  is.numeric(x) && length(x) == 1L && is.finite(x)
+  is.numeric(x) && length(x) == 1L && is.finite(x) && x >= lower && x <= upper
 }
 
 is_dateish <- function(x, null_ok = FALSE) {